requirements.txt
django==4.2.5
djangorestframework==3.14
drf_yasg==1.21.7
opencv-python==4.8.0.76
djangorestframework-simplejwt==5.3.0